requirements.txt
streamlit
torch
torchvision
pytorch-grad-cam
pillow
matplotlib
